Feature: Deleted the userPlayer and computerPlayer components as well as the Game component and consolidated them into the Gameboard and Player components.
